docker/backend/packages/phi-detector-remover/pyproject.toml
[project]
name = "phi-detector-remover"
version = "2.0.0"
description = "Modular PHI detection and removal for images using Presidio, LLMs, and Vision Models"
requires-python = ">=3.11"
dependencies = [
    # Core detection
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",

    # OCR
    "pytesseract>=0.3.8",

    # Image processing
    "opencv-python>=4.5.0",
    "Pillow>=9.0.0",

    # Data handling
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "polars>=0.20.0",

    # HTTP client for LLM/LVM APIs
    "httpx>=0.27.0",
]

[project.optional-dependencies]
# FastAPI web service
web = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
]

# LLM/LVM support (local inference)
llm = [
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "accelerate>=0.27.0",
]

# All optional features
all = [
    "phi-detector-remover[web,llm]",
]

# Development dependencies
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/phi_detector_remover"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
